Zmq: header only
