+++ fairmq/zeromq/FairMQTransportFactoryZMQ.h
@@ -16,8 +16,9 @@
 #define FAIRMQTRANSPORTFACTORYZMQ_H_
 
 #include <fairmq/zeromq/Context.h>
+#include <fairmq/Tools.h>
 #include <fairmq/ProgOptions.h>
-#include "FairMQTransportFactory.h"
+#include <FairMQTransportFactory.h>
 #include "FairMQMessageZMQ.h"
 #include "FairMQSocketZMQ.h"
 #include "FairMQPollerZMQ.h"
@@ -30,26 +31,47 @@
 class FairMQTransportFactoryZMQ final : public FairMQTransportFactory
 {
   public:
-    FairMQTransportFactoryZMQ(const std::string& id = "", const fair::mq::ProgOptions* config = nullptr);
+    FairMQTransportFactoryZMQ(const std::string& id = "", const fair::mq::ProgOptions* config = nullptr)
+        : FairMQTransportFactory(id)
+        , fCtx(nullptr)
+    {
+        int major, minor, patch;
+        zmq_version(&major, &minor, &patch);
+        LOG(debug) << "Transport: Using ZeroMQ library, version: " << major << "." << minor << "." << patch;
+
+        if (config) {
+            fCtx = fair::mq::tools::make_unique<fair::mq::zmq::Context>(config->GetProperty<int>("io-threads", 1));
+        } else {
+            LOG(debug) << "fair::mq::ProgOptions not available! Using defaults.";
+            fCtx = fair::mq::tools::make_unique<fair::mq::zmq::Context>(1);
+        }
+    }
+
     FairMQTransportFactoryZMQ(const FairMQTransportFactoryZMQ&) = delete;
     FairMQTransportFactoryZMQ operator=(const FairMQTransportFactoryZMQ&) = delete;
 
-    FairMQMessagePtr CreateMessage() override;
-    FairMQMessagePtr CreateMessage(const size_t size) override;
-    FairMQMessagePtr CreateMessage(void* data, const size_t size, fairmq_free_fn* ffn, void* hint = nullptr) override;
-    FairMQMessagePtr CreateMessage(FairMQUnmanagedRegionPtr& region, void* data, const size_t size, void* hint = 0) override;
+    FairMQMessagePtr CreateMessage() override { return fair::mq::tools::make_unique<FairMQMessageZMQ>(this); }
+    FairMQMessagePtr CreateMessage(const size_t size) override { return fair::mq::tools::make_unique<FairMQMessageZMQ>(size, this); }
+    FairMQMessagePtr CreateMessage(void* data, const size_t size, fairmq_free_fn* ffn, void* hint = nullptr) override { return fair::mq::tools::make_unique<FairMQMessageZMQ>(data, size, ffn, hint, this); }
+    FairMQMessagePtr CreateMessage(FairMQUnmanagedRegionPtr& region, void* data, const size_t size, void* hint = 0) override { return fair::mq::tools::make_unique<FairMQMessageZMQ>(region, data, size, hint, this); }
 
-    FairMQSocketPtr CreateSocket(const std::string& type, const std::string& name) override;
+    FairMQSocketPtr CreateSocket(const std::string& type, const std::string& name) override { return fair::mq::tools::make_unique<FairMQSocketZMQ>(*fCtx, type, name, GetId(), this); }
 
-    FairMQPollerPtr CreatePoller(const std::vector<FairMQChannel>& channels) const override;
-    FairMQPollerPtr CreatePoller(const std::vector<FairMQChannel*>& channels) const override;
-    FairMQPollerPtr CreatePoller(const std::unordered_map<std::string, std::vector<FairMQChannel>>& channelsMap, const std::vector<std::string>& channelList) const override;
+    FairMQPollerPtr CreatePoller(const std::vector<FairMQChannel>& channels) const override { return fair::mq::tools::make_unique<FairMQPollerZMQ>(channels); }
+    FairMQPollerPtr CreatePoller(const std::vector<FairMQChannel*>& channels) const override { return fair::mq::tools::make_unique<FairMQPollerZMQ>(channels); }
+    FairMQPollerPtr CreatePoller(const std::unordered_map<std::string, std::vector<FairMQChannel>>& channelsMap, const std::vector<std::string>& channelList) const override { return fair::mq::tools::make_unique<FairMQPollerZMQ>(channelsMap, channelList); }
 
-    FairMQUnmanagedRegionPtr CreateUnmanagedRegion(const size_t size, FairMQRegionCallback callback, const std::string& path = "", int flags = 0) override;
-    FairMQUnmanagedRegionPtr CreateUnmanagedRegion(const size_t size, FairMQRegionBulkCallback bulkCallback, const std::string& path = "", int flags = 0) override;
-    FairMQUnmanagedRegionPtr CreateUnmanagedRegion(const size_t size, const int64_t userFlags, FairMQRegionCallback callback, const std::string& path = "", int flags = 0) override;
-    FairMQUnmanagedRegionPtr CreateUnmanagedRegion(const size_t size, const int64_t userFlags, FairMQRegionBulkCallback bulkCallback, const std::string& path = "", int flags = 0) override;
-    FairMQUnmanagedRegionPtr CreateUnmanagedRegion(const size_t size, const int64_t userFlags, FairMQRegionCallback callback, FairMQRegionBulkCallback bulkCallback, const std::string& path = "", int flags = 0);
+    FairMQUnmanagedRegionPtr CreateUnmanagedRegion(const size_t size, FairMQRegionCallback callback, const std::string& path = "", int flags = 0) override { return CreateUnmanagedRegion(size, 0, callback, nullptr, path, flags); }
+    FairMQUnmanagedRegionPtr CreateUnmanagedRegion(const size_t size, FairMQRegionBulkCallback bulkCallback, const std::string& path = "", int flags = 0) override { return CreateUnmanagedRegion(size, 0, nullptr, bulkCallback, path, flags); }
+    FairMQUnmanagedRegionPtr CreateUnmanagedRegion(const size_t size, const int64_t userFlags, FairMQRegionCallback callback, const std::string& path = "", int flags = 0) override { return CreateUnmanagedRegion(size, userFlags, callback, nullptr, path, flags); }
+    FairMQUnmanagedRegionPtr CreateUnmanagedRegion(const size_t size, const int64_t userFlags, FairMQRegionBulkCallback bulkCallback, const std::string& path = "", int flags = 0) override { return CreateUnmanagedRegion(size, userFlags, nullptr, bulkCallback, path, flags); }
+    FairMQUnmanagedRegionPtr CreateUnmanagedRegion(const size_t size, const int64_t userFlags, FairMQRegionCallback callback, FairMQRegionBulkCallback bulkCallback, const std::string& path = "", int flags = 0)
+    {
+        auto ptr = std::unique_ptr<FairMQUnmanagedRegion>(new FairMQUnmanagedRegionZMQ(*fCtx, size, userFlags, callback, bulkCallback, path, flags, this));
+        auto zPtr = static_cast<FairMQUnmanagedRegionZMQ*>(ptr.get());
+        fCtx->AddRegion(zPtr->GetId(), zPtr->GetData(), zPtr->GetSize(), zPtr->GetUserFlags(), fair::mq::RegionEvent::created);
+        return ptr;
+    }
 
     void SubscribeToRegionEvents(FairMQRegionEventCallback callback) override { fCtx->SubscribeToRegionEvents(callback); }
     bool SubscribedToRegionEvents() override { return fCtx->SubscribedToRegionEvents(); }
@@ -62,7 +84,7 @@ class FairMQTransportFactoryZMQ final : public FairMQTransportFactory
     void Resume() override { fCtx->Resume(); }
     void Reset() override { fCtx->Reset(); }
 
-    ~FairMQTransportFactoryZMQ() override;
+    ~FairMQTransportFactoryZMQ() override { LOG(debug) << "Destroying ZeroMQ transport..."; }
 
   private:
     std::unique_ptr<fair::mq::zmq::Context> fCtx;
